% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Creates a manhattan plot.}
\usage{
manhattan(x, snp = "SNP", chr = "CHR", bp = "BP", p = "P",
  maf = "MAF", typed = NA, annotation = NA, thresholdLow = 5,
  thresholdHigh = -log10(5e-08), thresholdLowColor = "blue",
  thresholdHighColor = "red", mafColor = "black", build = "b37",
  title = Sys.time())
}
\arguments{
\item{x}{A data frame with result data}

\item{snp}{SNP column in data frame}

\item{chr}{Chromosome column in data frame}

\item{p}{P-value column in data frame}

\item{maf}{MAF column in data frame}

\item{typed}{the column in the data frame indicating whether the markers are genotyped or imputed (ignored if NA)}

\item{annotation}{a vector of annotation (ignored if NA)}

\item{thresholdLow}{the low threshold value (log10)}

\item{thresholdHigh}{the high threshold value (log10)}

\item{thresholdLowColor}{the color of the low threshold}

\item{thresholdHighColor}{the color of the high threshold}

\item{mafColor}{the color of the low maf values}

\item{build}{What build to use for plotting ('b37' or 'b38', default is 'b37')}

\item{title}{Title of plot (date by default, ignored if NA)}

\item{pb}{SNP position column in data frame}

\item{highlight}{Whether to highlight SNPs (default to FALSE)}

\item{highlight.col}{Vector containing SNPs to label}
}
\value{
A manhattan plot (ggplot2 object)
}
\description{
Creates a manhattan plot.
}
