% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.R
\name{qq}
\alias{qq}
\title{Creates a QQ plot using ggplot2 and returns the plot object.}
\usage{
qq(x, p = "P", maf = "MAF", typed = NA, thresholdLow = 5,
  thresholdHigh = -log10(5e-08), thresholdLowColor = "blue",
  thresholdHighColor = "red", mafColor = "black", title = Sys.time())
}
\arguments{
\item{x}{A data frame with result data}

\item{p}{P-value column in data frame}

\item{typed}{the column in the data frame indicating whether the markers are genotyped or imputed (ignored if NA)}

\item{thresholdLow}{the low threshold value (log10)}

\item{thresholdHigh}{the high threshold value (log10)}

\item{thresholdLowColor}{the color of the low threshold}

\item{thresholdHighColor}{the color of the high threshold}

\item{mafColor}{the color of the low maf values}

\item{title}{Title of plot (date by default, ignored if NA)}
}
\value{
A manhattan plot (ggplot2 object)
}
\description{
Creates a QQ plot using ggplot2 and returns the plot object.
}
